requirements.txt
Django>=5.2.5
Pillow>=11.3.0
gunicorn>=22.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
python-decouple>=3.8
